# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Flash Sale Management",
            "email": "jerdem.akyildiz@gmail.com"
        }
    },
    "paths": {
        "/flash-sales": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Get Flash All Sale",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.SaleResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Update Flash Sale",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateSaleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/response.SaleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Create Flash Sale",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateSaleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SaleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flash-sales/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Get Flash Sale",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flash Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/response.SaleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Delete Flash Sale",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flash Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/flash-sales/{id}/buy": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Buy Product",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateSaleRequest": {
            "type": "object",
            "required": [
                "discount",
                "endTime",
                "product_id",
                "saleStock",
                "startTime"
            ],
            "properties": {
                "discount": {
                    "type": "number"
                },
                "endTime": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "saleStock": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "request.UpdateSaleRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "discount": {
                    "type": "number"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "saleStock": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "response.SaleResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "discount": {
                    "type": "number"
                },
                "endTime": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "saleStock": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                }
            }
        }
    }
}